Fix C++20 errors on macOS
Fix deprecated-anon-enum-enum-conversion warning.
Fix aggregate initialization for C++20 semantics.
Add placeholder documentation for the C++ version we use.

+++ common/test/eigen_types_test.cc
@@ -220,8 +220,10 @@ GTEST_TEST(EigenTypesTest, FixedSizeVector) {
 GTEST_TEST(EigenTypesTest, LikewiseStorageOrder) {
   using LikewiseCol = MatrixLikewise<float, Vector3<double>>;
   using LikewiseRow = MatrixLikewise<float, RowVector3<double>>;
-  EXPECT_EQ(LikewiseCol::Options & Eigen::ColMajor, Eigen::ColMajor);
-  EXPECT_EQ(LikewiseRow::Options & Eigen::RowMajor, Eigen::RowMajor);
+  EXPECT_EQ(static_cast<int>(LikewiseCol::Options) & Eigen::ColMajor,
+            Eigen::ColMajor);
+  EXPECT_EQ(static_cast<int>(LikewiseRow::Options) & Eigen::RowMajor,
+            Eigen::RowMajor);
 }
 
 }  // namespace
